refactor: group personmodal and form
